refactor: Fix compiler diagnostics

+++ include/model/config/configuration.h
@@ -23,7 +23,7 @@ struct CompilationConfig {
 // 数据库相关配置
 struct DatabaseConfig {
   std::string path;
-  int batch_size;
+  size_t batch_size;
   int cache_size_mb;
   std::string journal_mode;
   std::string synchronous;
